Fix build on x86_64
A pointer does not equal to an int on x86_64
(which results into pointer - pointer not being an int either),
fix that by casting the resulting value  to an int.

Breakage was introduced by commit 909b5ec43c6f5ce2f97c5860b983610ae3d6d053

+++ src/st/st-widget.c
@@ -1514,7 +1514,7 @@ remove_class_name (gchar       **class_list,
   else
     end++;
 
-  new_class_list = g_strdup_printf ("%.*s%s", match - *class_list,
+  new_class_list = g_strdup_printf ("%.*s%s", (int)(match - *class_list),
                                     *class_list, end);
   g_free (*class_list);
   *class_list = new_class_list;
